SS4.0: Added ACL to Backend
git-svn-id: http://joomlacode.org/svn/sermon_speaker@250 5d6a53e0-5f49-0410-96cc-992822d07570

+++ SermonSpeaker4.0/com_sermonspeaker/admin/views/series/view.html.php
@@ -33,29 +33,40 @@ class SermonspeakerViewSeries extends JView
 	 */
 	protected function addToolbar()
 	{
-		$state	= $this->get('State');
+		$canDo 	= SermonspeakerHelper::getActions();
 
 		JToolBarHelper::title(JText::_('COM_SERMONSPEAKER_SERIES_TITLE'), 'series');
-		JToolBarHelper::addNew('serie.add','JTOOLBAR_NEW');
-		JToolBarHelper::editList('serie.edit','JTOOLBAR_EDIT');
-		JToolBarHelper::divider();
-		JToolBarHelper::custom('series.publish', 'publish.png', 'publish_f2.png','JTOOLBAR_PUBLISH', true);
-		JToolBarHelper::custom('series.unpublish', 'unpublish.png', 'unpublish_f2.png', 'JTOOLBAR_UNPUBLISH', true);
-		if ($state->get('filter.state') != -1 ) {
+
+		if ($canDo->get('core.create')) {
+			JToolBarHelper::addNew('sermon.add','JTOOLBAR_NEW');
+		}
+
+		if (($canDo->get('core.edit')) || ($canDo->get('core.edit.own'))) {
+			JToolBarHelper::editList('sermon.edit','JTOOLBAR_EDIT');
+		}
+
+		if ($canDo->get('core.edit.state')) {
 			JToolBarHelper::divider();
-			if ($state->get('filter.state') != 2) {
+			JToolBarHelper::custom('series.publish', 'publish.png', 'publish_f2.png','JTOOLBAR_PUBLISH', true);
+			JToolBarHelper::custom('series.unpublish', 'unpublish.png', 'unpublish_f2.png', 'JTOOLBAR_UNPUBLISH', true);
+			JToolBarHelper::divider();
+			if ($this->state->get('filter.state') != 2) {
 				JToolBarHelper::archiveList('series.archive','JTOOLBAR_ARCHIVE');
-			}
-			else if ($state->get('filter.state') == 2) {
+			} else {
 				JToolBarHelper::unarchiveList('series.publish', 'JTOOLBAR_UNARCHIVE');
 			}
 		}
-		if ($state->get('filter.state') == -2) {
+
+		if ($this->state->get('filter.state') == -2 && $canDo->get('core.delete')) {
 			JToolBarHelper::deleteList('', 'series.delete','JTOOLBAR_EMPTY_TRASH');
-		} else {
+			JToolBarHelper::divider();
+		} else if ($canDo->get('core.edit.state')) {
 			JToolBarHelper::trash('series.trash','JTOOLBAR_TRASH');
+			JToolBarHelper::divider();
 		}
-		JToolBarHelper::divider();
-		JToolBarHelper::preferences('com_sermonspeaker');
+
+		if ($canDo->get('core.admin')) {
+			JToolBarHelper::preferences('com_sermonspeaker', 550, 800);
+		}
 	}
 }
